refactor: re-organize array observation code
