+++ packages/web-components/fast-element/src/observation/splice-strategies.ts
@@ -1,6 +1,10 @@
 import { emptyArray } from "../platform.js";
-import type { ArrayObserver } from "./array-observer.js";
-import { Splice, SpliceStrategy, SpliceStrategySupport } from "./splice.js";
+import {
+    ArrayObserver,
+    Splice,
+    SpliceStrategy,
+    SpliceStrategySupport,
+} from "./arrays.js";
 
 const enum Edit {
     leave = 0,
@@ -394,53 +398,122 @@ function project(array: unknown[], changes: Splice[]): Splice[] {
  * splices needed to represent the change from the old array to the new array.
  * @public
  */
-export const mergeSpliceStrategy: SpliceStrategy = Object.freeze(
-    Object.assign({}, SpliceStrategy.default, {
-        support: SpliceStrategySupport.optimized,
+export const mergeSpliceStrategy: SpliceStrategy = Object.freeze({
+    support: SpliceStrategySupport.optimized,
 
-        normalize(
-            previous: unknown[] | undefined,
-            current: unknown[],
-            changes: Splice[] | undefined
-        ): readonly Splice[] {
-            if (previous === void 0) {
-                if (changes === void 0) {
-                    return emptyArray;
-                }
-
-                return changes.length > 1 ? project(current, changes) : changes;
+    normalize(
+        previous: unknown[] | undefined,
+        current: unknown[],
+        changes: Splice[] | undefined
+    ): readonly Splice[] {
+        if (previous === void 0) {
+            if (changes === void 0) {
+                return emptyArray;
             }
 
-            return calc(current, 0, current.length, previous, 0, previous.length);
-        },
+            return changes.length > 1 ? project(current, changes) : changes;
+        }
 
-        reverse(
-            array: any[],
-            observer: ArrayObserver,
-            reverse: typeof Array.prototype.reverse,
-            args: any[]
-        ): any {
-            observer.flush();
-            const oldArray = array.slice();
-            const result = reverse.apply(array, args);
-            observer.reset(oldArray);
-            return result;
-        },
+        return calc(current, 0, current.length, previous, 0, previous.length);
+    },
 
-        sort(
-            array: any[],
-            observer: ArrayObserver,
-            sort: typeof Array.prototype.sort,
-            args: any[]
-        ): any[] {
-            observer.flush();
-            const oldArray = array.slice();
-            const result = sort.apply(array, args);
-            observer.reset(oldArray);
-            return result;
-        },
-    })
-);
+    pop(
+        array: any[],
+        observer: ArrayObserver,
+        pop: typeof Array.prototype.pop,
+        args: any[]
+    ) {
+        const notEmpty = array.length > 0;
+        const result = pop.apply(array, args);
+
+        if (notEmpty) {
+            observer.addSplice(new Splice(array.length, [result], 0));
+        }
+
+        return result;
+    },
+
+    push(
+        array: any[],
+        observer: ArrayObserver,
+        push: typeof Array.prototype.push,
+        args: any[]
+    ): any {
+        const result = push.apply(array, args);
+        observer.addSplice(
+            new Splice(array.length - args.length, [], args.length).adjustTo(array)
+        );
+        return result;
+    },
+
+    reverse(
+        array: any[],
+        observer: ArrayObserver,
+        reverse: typeof Array.prototype.reverse,
+        args: any[]
+    ): any {
+        observer.flush();
+        const oldArray = array.slice();
+        const result = reverse.apply(array, args);
+        observer.reset(oldArray);
+        return result;
+    },
+
+    shift(
+        array: any[],
+        observer: ArrayObserver,
+        shift: typeof Array.prototype.shift,
+        args: any[]
+    ): any {
+        const notEmpty = array.length > 0;
+        const result = shift.apply(array, args);
+
+        if (notEmpty) {
+            observer.addSplice(new Splice(0, [result], 0));
+        }
+
+        return result;
+    },
+
+    sort(
+        array: any[],
+        observer: ArrayObserver,
+        sort: typeof Array.prototype.sort,
+        args: any[]
+    ): any[] {
+        observer.flush();
+        const oldArray = array.slice();
+        const result = sort.apply(array, args);
+        observer.reset(oldArray);
+        return result;
+    },
+
+    splice(
+        array: any[],
+        observer: ArrayObserver,
+        splice: typeof Array.prototype.splice,
+        args: any[]
+    ): any {
+        const result = splice.apply(array, args);
+        observer.addSplice(
+            new Splice(+args[0], result, args.length > 2 ? args.length - 2 : 0).adjustTo(
+                array
+            )
+        );
+        return result;
+    },
+
+    unshift(
+        array: any[],
+        observer: ArrayObserver,
+        unshift: typeof Array.prototype.unshift,
+        args: any[]
+    ): any[] {
+        const result = unshift.apply(array, args);
+        observer.addSplice(new Splice(0, [], args.length).adjustTo(array));
+        return result;
+    },
+});
 
 /**
  * A splice strategy that doesn't create splices, but instead
